migrator object
scooped up all the logic from self.migrate into a Migrator object. might
be better termed a Migration? at any rate, I need to see how closely this
comes to working before I go any further, but if this commit is in the
log, then this approach worked. probably (in that case) I'll build on this
with the ability to override it with subclasses, and keep the methods you
want.